+++ NEWS.md
@@ -17,6 +17,7 @@
     can also parse by their simple name instead of requiring a complex object.
   - Comictagger schema supports series_aliases and title_aliases as reprints
   - CB7 archive read support. Comic archives in 7zip archives.
+  - Syntax highlighting used on output. Change or remove with --theme option.
 
 - Fixes
   - ComicInfo.xml StoryArcs tag was not loaded.